Make the implementation conform to interface
We now panic on invalid input.

This change required a change to cblas.h since it incorrectly named Ap
as A in some signatures, so genBlas was confused.

This change requires testing: Order may interact with tests in a way
that I have not understood (all the assertions were gleaned from reading
netlib .f files, so where the documentation is unclear I have examined
the fortran - that code doesn't test array lengths, so that aspect of
this change may be wrong).