Fix incorrect parameter checks
Testing the previous changes in the context of biogo.blas and
biogo.matrix identified a number of incorrect checks - mainly due to the
Order parameter, which can not be understood purely with reference to
the fortran code. Reference to the ATLAS C wrappers has clarified this
to a degree, so it should be OK. biogo.matrix tests pass with this code,
but further testing is required.